Preparing to revamp imagePickerController. Colin says to just addsubview of a button that allow on click, changes the sourcetype from camera to photo library (and reverse).
Also considering his (and Brian's) advice to change main UI to be more iOS7-y. As in a white background instead of blue and better colored buttons. Or instead, change the button colors and button background to be easier to read. Not sure what to do for UI changes yet.

Either way, will need to do UIImagePickerController first, and then have Picker horizontal implemented for icons in place of the current add image. Also will need to consider whether or not to add a Navigation Bar at the beginning (see previous paragraph for more UI)